+++ lib/cvloader.ts
@@ -88,7 +88,7 @@ function getOpenCV(opt?: OpenCVBuildEnvParams): OpenCVType {
           Failed with: ${pc.red(message)}, openCV binding not available, reed: 
           build-opencv --help
           And build missing file with:
-          npx build-opencv --version 4.5.4 rebuild
+          npx build-opencv --version 4.6.0 rebuild
           
           PS: a 'npm link' may help
           `;
@@ -96,7 +96,7 @@ function getOpenCV(opt?: OpenCVBuildEnvParams): OpenCVType {
           msg = `require("${pc.yellow(requirePath)}"); 
           Failed with: ${pc.red(message)}, openCV module looks broken, clean you builds directory and rebuild everything
           rm -r <path to your build directory>
-          npx build-opencv --version 4.5.4 rebuild
+          npx build-opencv --version 4.6.0 rebuild
           `;
         } else {
           msg = `require("${pc.yellow(requirePath)}");